# docs/swagger.yaml
basePath: /
definitions:
...
  common.BasicResponse:
    properties:
      message:
        type: string
    type: object
...
  common.GeneralResponse:
    properties:
      message:
        type: string
      result: {}
    type: object
  common.PaginatedResponse:
    properties:
      metadata:
        $ref: '#/definitions/pagination.Metadata'
      result:
        items: {}
        type: array
    type: object
  company.Company:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      region:
        type: string
      updatedAt:
        type: string
    type: object
  company.CompanyDTO:
    properties:
      name:
        type: string
      region:
        type: string
    required:
    - name
    - region
    type: object
  employee.Employee:
    properties:
      address:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      phonenumber:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  employee.EmployeeCreateDTO:
    properties:
      address:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      phonenumber:
        type: string
      status:
        type: boolean
      username:
        type: string
    required:
    - name
    - password
    - status
    - username
    type: object
  employee.EmployeeUpdateDTO:
    properties:
      address:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      phonenumber:
        type: string
      status:
        type: boolean
      username:
        type: string
    required:
    - name
    - status
    - username
    type: object
...
  pagination.Metadata:
    properties:
      count:
        type: integer
      hasNext:
        type: boolean
      hasPrev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  user.User:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        enum:
        - superadmin
        - owner
        - employee
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
info:
  contact:
    email: infamous0192@gmail.com
    name: Dwa Meizadewa
  description: Aplikasi Backend (API) untuk project Abude
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Abude Backend
  version: 2.0.0
paths:
...
  /api/company:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: keyword
        type: string
      - in: query
        name: limit
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: integer
        name: owner
        type: array
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/company.Company'
                  type: array
              type: object
      security:
      - JWT: []
      summary: Get All Company
      tags:
      - Companies
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/company.CompanyDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/company.Company'
              type: object
      security:
      - JWT: []
      summary: Create Company
      tags:
      - Companies
  /api/company/:id/owner:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/user.User'
                  type: array
              type: object
      security:
      - JWT: []
      summary: Get Company Owners
      tags:
      - Companies
  /api/company/{companyId}/owner/{userId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Company ID
        in: path
        name: companyId
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BasicResponse'
      security:
      - JWT: []
      summary: Remove Company Owner
      tags:
      - Companies
    put:
      consumes:
      - application/json
      parameters:
      - description: Company ID
        in: path
        name: companyId
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.BasicResponse'
      security:
      - JWT: []
      summary: Add Company Owner
      tags:
      - Companies
  /api/company/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/company.Company'
              type: object
      security:
      - JWT: []
      summary: Delete Company
      tags:
      - Companies
    get:
      consumes:
      - application/json
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/company.Company'
      security:
      - JWT: []
      summary: Get One Company
      tags:
      - Companies
    put:
      consumes:
      - application/json
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/company.CompanyDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/company.Company'
              type: object
      security:
      - JWT: []
      summary: Update Company
      tags:
      - Companies
  /api/employee:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: keyword
        type: string
      - in: query
        name: limit
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: integer
        name: outlet
        type: array
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/employee.Employee'
                  type: array
              type: object
      security:
      - JWT: []
      summary: Get All Employee
      tags:
      - Employees
    post:
      consumes:
      - application/json
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/employee.EmployeeCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/employee.Employee'
              type: object
      security:
      - JWT: []
      summary: Create Employee
      tags:
      - Employees
  /api/employee/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/employee.Employee'
              type: object
      security:
      - JWT: []
      summary: Delete Employee
      tags:
      - Employees
    get:
      consumes:
      - application/json
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/employee.Employee'
      security:
      - JWT: []
      summary: Get One Employee
      tags:
      - Employees
    put:
      consumes:
      - application/json
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/employee.EmployeeUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.GeneralResponse'
            - properties:
                result:
                  $ref: '#/definitions/employee.Employee'
              type: object
      security:
      - JWT: []
      summary: Update Employee
      tags:
      - Employees
...
schemes:
- https
- http
securityDefinitions:
  JWT:
    description: Authorization for JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"